Move rolie check code to respective file.
