+++ cmd/csaf_checker/processor.go
@@ -947,184 +947,6 @@ func (p *processor) checkChanges(base string, mask whereType) error {
 	return p.integrity(files, base, mask, p.badChanges.add)
 }
 
-func tlpLabel(label *csaf.TLPLabel) csaf.TLPLabel {
-	if label != nil {
-		return *label
-	}
-	return csaf.TLPLabelUnlabeled
-}
-
-func (p *processor) processROLIEFeeds(feeds [][]csaf.Feed) error {
-
-	base, err := url.Parse(p.pmdURL)
-	if err != nil {
-		return err
-	}
-	p.badROLIEfeed.use()
-
-	// Phase 1: load all advisories urls.
-
-	advisories := map[*csaf.Feed][]csaf.AdvisoryFile{}
-
-	for _, fs := range feeds {
-		for i := range fs {
-			feed := &fs[i]
-			if feed.URL == nil {
-				continue
-			}
-			up, err := url.Parse(string(*feed.URL))
-			if err != nil {
-				p.badProviderMetadata.error("Invalid URL %s in feed: %v.", *feed.URL, err)
-				continue
-			}
-			feedBase := base.ResolveReference(up)
-			feedURL := feedBase.String()
-			p.checkTLS(feedURL)
-
-			advs, err := p.rolieFeedEntries(feedURL)
-			if err != nil {
-				if err != errContinue {
-					return err
-				}
-				continue
-			}
-			advisories[feed] = advs
-		}
-	}
-
-	// Phase 2:
-
-	// Indicates if there is a complete feed for a given label.
-	for _, fs := range feeds {
-		for i := range fs {
-			feed := &fs[i]
-			if feed.URL == nil {
-				continue
-			}
-			files := advisories[feed]
-			if files == nil {
-				continue
-			}
-
-			up, err := url.Parse(string(*feed.URL))
-			if err != nil {
-				p.badProviderMetadata.error("Invalid URL %s in feed: %v.", *feed.URL, err)
-				continue
-			}
-
-			feedURL := base.ResolveReference(up)
-			feedBase, err := util.BaseURL(feedURL)
-			if err != nil {
-				p.badProviderMetadata.error("Bad base path: %v", err)
-				continue
-			}
-
-			label := tlpLabel(feed.TLPLabel)
-
-			p.labelChecker = &rolieLabelChecker{
-				feedURL:    feedURL.String(),
-				feedLabel:  label,
-				advisories: map[csaf.TLPLabel]map[string]struct{}{},
-			}
-
-			if err := p.integrity(files, feedBase, rolieMask, p.badProviderMetadata.add); err != nil {
-				if err != errContinue {
-					return err
-				}
-			}
-		}
-	}
-
-	// Phase 3: Check for completeness.
-
-	hasSummary := map[csaf.TLPLabel]struct{}{}
-
-	var (
-		hasUnlabeled = false
-		hasWhite     = false
-		hasGreen     = false
-	)
-
-	for _, fs := range feeds {
-		for i := range fs {
-			feed := &fs[i]
-			if feed.URL == nil {
-				continue
-			}
-			files := advisories[feed]
-			if files == nil {
-				continue
-			}
-
-			up, err := url.Parse(string(*feed.URL))
-			if err != nil {
-				p.badProviderMetadata.error("Invalid URL %s in feed: %v.", *feed.URL, err)
-				continue
-			}
-
-			feedBase := base.ResolveReference(up)
-			makeAbs := makeAbsolute(feedBase)
-			label := tlpLabel(feed.TLPLabel)
-
-			switch label {
-			case csaf.TLPLabelUnlabeled:
-				hasUnlabeled = true
-			case csaf.TLPLabelWhite:
-				hasWhite = true
-			case csaf.TLPLabelGreen:
-				hasGreen = true
-			}
-
-			reference := p.labelChecker.advisories[label]
-			advisories := make(map[string]struct{}, len(reference))
-
-			for _, adv := range files {
-				u, err := url.Parse(adv.URL())
-				if err != nil {
-					p.badProviderMetadata.error("Invalid URL %s in feed: %v.", *feed.URL, err)
-					continue
-				}
-				advisories[makeAbs(u).String()] = struct{}{}
-			}
-			if containsAllKeys(reference, advisories) {
-				hasSummary[label] = struct{}{}
-			}
-		}
-	}
-
-	if !hasWhite && !hasGreen && !hasUnlabeled {
-		p.badROLIEfeed.error(
-			"One ROLIE feed with a TLP:WHITE, TLP:GREEN or unlabeled tlp must exist, but none were found.")
-	}
-
-	// Every TLP level with data should have at least on summary feed.
-	for _, label := range []csaf.TLPLabel{
-		csaf.TLPLabelUnlabeled,
-		csaf.TLPLabelWhite,
-		csaf.TLPLabelGreen,
-		csaf.TLPLabelAmber,
-		csaf.TLPLabelRed,
-	} {
-		if _, ok := hasSummary[label]; !ok && len(p.labelChecker.advisories[label]) > 0 {
-			p.badROLIEfeed.warn(
-				"ROLIE feed for TLP:%s has no accessible listed feed covering all advisories.",
-				label)
-		}
-	}
-
-	return nil
-}
-
-// containsAllKeys returns if m2 contains all keys of m1.
-func containsAllKeys[K comparable, V any](m1, m2 map[K]V) bool {
-	for k := range m1 {
-		if _, ok := m2[k]; !ok {
-			return false
-		}
-	}
-	return true
-}
-
 // empty checks if list of strings contains at least one none empty string.
 func empty(arr []string) bool {
 	for _, s := range arr {
